+++ toys/builtins/system.rb
@@ -29,8 +29,7 @@
 
 desc "A group of system commands for toys"
 
-long_desc wrappable("The \"toys system\" group contains tools that inspect, configure, and" \
-                    " update toys itself.")
+long_desc "Contains tools that inspect, configure, and update toys itself."
 
 tool "version" do
   desc "Print current toys version."
@@ -43,8 +42,8 @@ end
 tool "update" do
   desc "Update toys if a newer version is available."
 
-  long_desc wrappable("Checks rubygems for a newer version of toys. If one is available," \
-                      " downloads and installs it.")
+  long_desc "Checks rubygems for a newer version of toys. If one is available, downloads" \
+            " and installs it."
 
   flag :yes, "-y", "--yes", desc: "Do not ask for interactive confirmation"